feat: 상품 생성 API 일부 코드 모듈화 제품 생성에 필요한 데이터를 준비하는 코드를 모듈화
